+++ src/main/java/com/pierrette/api/services/OperateurService.java
@@ -13,19 +13,38 @@ public class OperateurService {
     @Autowired
     private OperateurRepo operateurRepository;
 
-    public List<Operateur> findAll() {
-        return operateurRepository.findAll();
-    }
 
-    public Optional<Operateur> findById(Integer id) {
-        return operateurRepository.findById(id);
-    }
+    public List<Operateur> getAllOperateurs(){
+       return operateurRepository.findAll();
+   }
 
-    public Operateur save(Operateur operateur) {
+    public Operateur createOperateur(Operateur operateur){
         return operateurRepository.save(operateur);
+   }
+
+    public Operateur updateOperateur(Integer id, Operateur operateurUpdate){
+        Operateur operateurExiste = operateurRepository.findById(id).orElse(null);
+        if (operateurExiste!=null){
+            operateurExiste.setNom(operateurUpdate.getNom());
+            operateurExiste.setPrenom(operateurUpdate.getPrenom());
+            operateurExiste.setTelephone(operateurUpdate.getTelephone());
+            operateurExiste.setSexe(operateurUpdate.getSexe());
+
+            Operateur operateurAjour = operateurRepository.save(operateurExiste);
+
+        } return operateurExiste;
     }
 
-    public void deleteById(Integer id) {
-        operateurRepository.deleteById(id);
+    public String deleteOperateur(Integer id) {
+        if(id!=0){
+            operateurRepository.deleteById(id);
+            return "Suppression";
+        }
+        return null;
     }
+
+    public Operateur getOperateur(Integer idOperateur){
+        return operateurRepository.findById(idOperateur).get();
+   }
+
 }
