+++ src/main/java/com/pierrette/api/services/PaiementService.java
@@ -1,6 +1,8 @@
 package com.pierrette.api.services;
 
+import com.pierrette.api.entities.Operateur;
 import com.pierrette.api.entities.Paiement;
+import com.pierrette.api.repositories.OperateurRepo;
 import com.pierrette.api.repositories.PaiementRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,19 +15,36 @@ public class PaiementService {
     @Autowired
     private PaiementRepo paiementRepository;
 
-    public List<Paiement> findAll() {
+
+    public List<Paiement> getAllPaiements(){
         return paiementRepository.findAll();
     }
 
-    public Optional<Paiement> findById(Integer id) {
-        return paiementRepository.findById(id);
-    }
-
-    public Paiement save(Paiement paiement) {
+    public Paiement createPaiement(Paiement paiement){
         return paiementRepository.save(paiement);
     }
 
-    public void deleteById(Integer id) {
-        paiementRepository.deleteById(id);
+    public Paiement updatePaiement(Integer id, Paiement paiementUpdate){
+       Paiement paiementExiste = paiementRepository.findById(id).orElse(null);
+        if (paiementExiste!=null){
+            paiementExiste.setDatePaiement(paiementUpdate.getDatePaiement());
+            paiementExiste.setMontant(paiementUpdate.getMontant());
+
+           Paiement paiementAjour = paiementRepository.save(paiementExiste);
+
+        } return paiementExiste;
     }
+
+    public String deletePaiement(Integer id) {
+        if(id!=0){
+            paiementRepository.deleteById(id);
+            return "Suppression";
+        }
+        return null;
+    }
+
+    public Paiement getPaiement(Integer idPaiement){
+        return paiementRepository.findById(idPaiement).get();
+    }
+
 }
