+++ src/main/java/com/pierrette/api/services/PrefectureService.java
@@ -1,6 +1,8 @@
 package com.pierrette.api.services;
 
+import com.pierrette.api.entities.Operateur;
 import com.pierrette.api.entities.Prefecture;
+import com.pierrette.api.repositories.OperateurRepo;
 import com.pierrette.api.repositories.PrefectureRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -13,20 +15,34 @@ public class PrefectureService {
     @Autowired
     private PrefectureRepo prefectureRepository;
 
-    public List<Prefecture> findAll() {
+
+    public List<Prefecture> getAllPrefectures(){
         return prefectureRepository.findAll();
     }
 
-    public Optional<Prefecture> findById(Integer id) {
-        return prefectureRepository.findById(id);
-    }
-
-    public Prefecture save(Prefecture prefecture) {
+    public Prefecture createPrefecture(Prefecture prefecture){
         return prefectureRepository.save(prefecture);
     }
 
-    public void deleteById(Integer id) {
-        prefectureRepository.deleteById(id);
+    public Prefecture updatePrefecture(Integer id, Prefecture prefectureUpdate){
+        Prefecture prefectureExiste = prefectureRepository.findById(id).orElse(null);
+        if (prefectureExiste!=null){
+            prefectureExiste.setLibellePrefecture(prefectureUpdate.getLibellePrefecture());
+            prefectureExiste.setLibelleVille(prefectureUpdate.getLibelleVille());
+            Prefecture prefectureAjour = prefectureRepository.save(prefectureExiste);
+        } return prefectureExiste;
     }
-}
 
+    public String deletePrefecture(Integer id) {
+        if(id!=0){
+            prefectureRepository.deleteById(id);
+            return "Suppression";
+        }
+        return null;
+    }
+
+    public Prefecture getPrefecture(Integer idPrefecture){
+        return prefectureRepository.findById(idPrefecture).get();
+    }
+
+}
